Rename powf to pow (since it is different from glam's)

+++ src/sl/gen_type.rs
@@ -37,7 +37,7 @@ macro_rules! impl_gen_type {
                 super::primitives::built_in_2("atan", self, x.to_value())
             }
 
-            pub fn powf(self, y: impl super::ToValue<Output = Self>) -> Self {
+            pub fn pow(self, y: impl super::ToValue<Output = Self>) -> Self {
                 super::primitives::built_in_2("pow", self, y.to_value())
             }
